# Vectorize the softmax calculation when not along the last dim