--- a/aten/src/ATen/cpu/vec/vec256/vec256_bfloat16.h
+++ b/aten/src/ATen/cpu/vec/vec256/vec256_bfloat16.h
@@ -703,16 +703,6 @@ Vectorized<BFloat16> inline fmadd(const Vectorized<BFloat16>& a,
   return cvtfp32_bf16(o1, o2);
 }
 
-inline std::tuple<Vectorized<float>, Vectorized<float>> convert_bfloat16_float(const Vectorized<BFloat16>& a) {
-  __m256 o1, o2;
-  cvtbf16_fp32(__m256i(a), o1, o2);
-  return std::make_tuple(o1, o2);
-}
-
-inline Vectorized<BFloat16> convert_float_bfloat16(const Vectorized<float>& a, const Vectorized<float>& b) {
- return cvtfp32_bf16(__m256(a), __m256(b));
-}
-
 #endif
 
 }}}