@leslie-fang-intel 2021-05-31
Hi @yf225, I have met some `TestCppApiParity` UT fails in this PR as  
```
test_torch_nn_Softmax2d - TestCppApiParity
test_cpp_api_parity.py
```
However, I can pass the test in my local env with command: `python test/test_cpp_api_parity.py TestCppApiParity.test_torch_nn_Softmax2d`

Do you know how can I reproduce the fails in the UT?