@leslie-fang-intel 2021-06-02
> Hello @leslie-fang-intel, thank you for your PR! Could you also post some benchmarks for other (smaller and larger) shapes and dimensions including benchmarks on some different dtypes that are also being implemented here? Also, how does this scale with the number of CPU cores available? Thank you.

Hi @cpuhrsch, thanks for the suggestions. The [original implementation](https://github.com/pytorch/pytorch/blob/d417a094f398f1c4efd7f818b14b8471a597fbcc/aten/src/ATen/native/SoftMax.cpp#L143-L145) only support dtype with fp32 and fp64. So for this PR, I think we should keep within support of these 2 dtypes. And as for the extra support of BF16, I can open up a following PR if this one looks ok.

Here is the performance data I collect on 28cores' Intel 8380H CPU with different shapes and core numbers:




<meta name="ProgId" content="Excel.Sheet">
<meta name="Generator" content="Microsoft Excel 15">
<link id="Main-File" rel="Main-File" href="file:///C:/Users/jiahaofa/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel="File-List" href="file:///C:/Users/jiahaofa/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<style>
<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
-->
</style>





Shape | Dtype | Dim | Cores | Base Performance(ms) | Optimized Performance(ms)
-- | -- | -- | -- | -- | --
[32, 81, 15130] | fp32 | -2 | 28 | 23.7605 | 8.6088
[32, 16, 15130] | fp32 | -2 | 28 | 3.8581 | 0.8681
[1, 81, 15130] | fp32 | -2 | 28 | 0.5321 | 0.0821
[1, 16, 15130] | fp32 | -2 | 28 | 0.1178 | 0.0221
[32, 81, 20] | fp32 | -2 | 28 | 0.0316 | 0.0261
[32, 16, 20] | fp32 | -2 | 28 | 0.0261 | 0.0265
[1, 81, 20] | fp32 | -2 | 28 | 0.0409 | 0.0366
[1, 16, 20] | fp32 | -2 | 28 | 0.0137 | 0.0129
[32, 81, 16] | fp32 | -2 | 28 | 0.0296 | 0.0274
[32, 16, 16] | fp32 | -2 | 28 | 0.015 | 0.0118
[1, 81, 16] | fp32 | -2 | 28 | 0.034 | 0.0319
[1, 16, 16] | fp32 | -2 | 28 | 0.0121 | 0.0117
[32, 81, 3] | fp32 | -2 | 28 | 0.0113 | 0.0107
[32, 16, 3] | fp32 | -2 | 28 | 0.0086 | 0.0078
[1, 81, 3] | fp32 | -2 | 28 | 0.0083 | 0.0078
[1, 16, 3] | fp32 | -2 | 28 | 0.0072 | 0.0071
[32, 81, 15130] | fp32 | -2 | 1 | 463.0909 | 115.262
[32, 16, 15130] | fp32 | -2 | 1 | 88.8961 | 12.469
[1, 81, 15130] | fp32 | -2 | 1 | 12.8996 | 1.7655
[1, 16, 15130] | fp32 | -2 | 1 | 2.5987 | 0.3342
[32, 81, 20] | fp32 | -2 | 1 | 0.5092 | 0.1445
[32, 16, 20] | fp32 | -2 | 1 | 0.1215 | 0.0437
[1, 81, 20] | fp32 | -2 | 1 | 0.0182 | 0.0069
[1, 16, 20] | fp32 | -2 | 1 | 0.0056 | 0.0034
[32, 81, 16] | fp32 | -2 | 1 | 0.4094 | 0.0541
[32, 16, 16] | fp32 | -2 | 1 | 0.0894 | 0.0142
[1, 81, 16] | fp32 | -2 | 1 | 0.0151 | 0.0039
[1, 16, 16] | fp32 | -2 | 1 | 0.0051 | 0.0027
[32, 81, 3] | fp32 | -2 | 1 | 0.0781 | 0.0693
[32, 16, 3] | fp32 | -2 | 1 | 0.018 | 0.0193
[1, 81, 3] | fp32 | -2 | 1 | 0.0047 | 0.0046
[1, 16, 3] | fp32 | -2 | 1 | 0.0027 | 0.0029
[32, 81, 15130] | fp64 | -2 | 28 | 40.4192 | 15.3224
[32, 16, 15130] | fp64 | -2 | 28 | 8.9532 | 3.1854
[1, 81, 15130] | fp64 | -2 | 28 | 3.1217 | 0.4471
[1, 16, 15130] | fp64 | -2 | 28 | 0.1816 | 0.0261
[32, 81, 20] | fp64 | -2 | 28 | 0.0395 | 0.0271
[32, 16, 20] | fp64 | -2 | 28 | 0.0139 | 0.0211
[1, 81, 20] | fp64 | -2 | 28 | 0.0317 | 0.0368
[1, 16, 20] | fp64 | -2 | 28 | 0.0126 | 0.0139
[32, 81, 16] | fp64 | -2 | 28 | 0.0341 | 0.032
[32, 16, 16] | fp64 | -2 | 28 | 0.0132 | 0.013
[1, 81, 16] | fp64 | -2 | 28 | 0.0255 | 0.0303
[1, 16, 16] | fp64 | -2 | 28 | 0.0114 | 0.0132
[32, 81, 3] | fp64 | -2 | 28 | 0.0126 | 0.0144
[32, 16, 3] | fp64 | -2 | 28 | 0.0083 | 0.0103
[1, 81, 3] | fp64 | -2 | 28 | 0.0088 | 0.0102
[1, 16, 3] | fp64 | -2 | 28 | 0.0068 | 0.0088
[32, 81, 15130] | fp64 | -2 | 1 | 724.4058 | 172.5648
[32, 16, 15130] | fp64 | -2 | 1 | 143.4831 | 26.8466
[1, 81, 15130] | fp64 | -2 | 1 | 18.7121 | 2.4747
[1, 16, 15130] | fp64 | -2 | 1 | 3.7613 | 0.4123
[32, 81, 20] | fp64 | -2 | 1 | 0.7511 | 0.2147
[32, 16, 20] | fp64 | -2 | 1 | 0.1659 | 0.0556
[1, 81, 20] | fp64 | -2 | 1 | 0.0271 | 0.0089
[1, 16, 20] | fp64 | -2 | 1 | 0.0077 | 0.0038
[32, 81, 16] | fp64 | -2 | 1 | 0.6073 | 0.0678
[32, 16, 16] | fp64 | -2 | 1 | 0.1232 | 0.0155
[1, 81, 16] | fp64 | -2 | 1 | 0.021 | 0.0043
[1, 16, 16] | fp64 | -2 | 1 | 0.0061 | 0.0027
[32, 81, 3] | fp64 | -2 | 1 | 0.1111 | 0.1089
[32, 16, 3] | fp64 | -2 | 1 | 0.025 | 0.0249
[1, 81, 3] | fp64 | -2 | 1 | 0.0057 | 0.0055
[1, 16, 3] | fp64 | -2 | 1 | 0.0028 | 0.0028