@cpuhrsch 2021-06-03
@leslie-fang-intel - thanks, those numbers look great!

The [test failures](https://app.circleci.com/pipelines/github/pytorch/pytorch/329728/workflows/4e9abf12-3c0b-4e4b-95f7-08c413ef3b07/jobs/13833140/tests#failed-test-0) seem legitimate, in particular I'm seeing

```
test_torch_nn_Softmax2d
test_torch_nn_Softmin_multidim
test_torch_nn_functional_softmax_functional_dim0
test_torch_nn_functional_softmax_spatial
test_torch_nn_functional_softmax_spatial_dtype
```

fail with a few nan comparisons.