@leslie-fang-intel 2021-06-04
@cpuhrsch I think most of the failures has been resolved. The left 2 failures may relate with the code base. I will rebase and try again. And I have updated the performance data in above tables with dtype of double.